open the articles urls inside a dialog (webview)
- When using webview by clicking 'Read more' article's link will be
loaded inside a dialog

- A progress bar will be shown on page started loading

- The progress bar will stop on page finished loading

- Added a Switch button to enable Javascript without going to settings

- WebView do not need app's restart since there is no JS option to be
shown

Why?

Because

1) WebView shows problems when a theme is set. This way only the dialog
will follow app's theme and the content will load following the native
webpage.

2) This loading's method is faster than removing views programmatically
land show the webview on top of another layout to be removed (like in
the past). WebView will be shown only when page loading is finished and
only progress bar will be hidden programmatically.